#2 Update tests to use latest version of php-markdown
